fix(loader_runner.py): change LAUNCH_METHOD value from "locally" to "cloud_manager" to improve semantics fix(remote_machine.py): change assert statement to check for "cloud_manager" as launch_method instead of "remotely" to improve semantics fix(script_execution.py): add conditional check to close zellij tab if launch_method is "cloud_manager" to improve functionality feat(session_managers.py): add close_tab method to close zellij tab by name in a zellij session

+++ myresources/crocodile/cluster/script_execution.py
@@ -156,6 +156,11 @@ rm_conf: RemoteMachineConfig = tb.Read.vanilla_pickle(path=manager.remote_machin
 
 
 if rm_conf.kill_on_completion and params.session_name != "":
-    from machineconfig.utils.procs import ProcessManager
-    pm = ProcessManager()
-    pm.kill(commands=[params.session_name])
+    if rm_conf.launch_method == "cloud_manager":
+        from crocodile.cluster.session_managers import Zellij
+        current_zellij_session = Zellij.get_current_zellij_session()
+        Zellij.close_tab(session_name=current_zellij_session, tab_name=params.session_name)
+    else:
+        from machineconfig.utils.procs import ProcessManager
+        pm = ProcessManager()
+        pm.kill(commands=[params.session_name])
